pyproject.toml
[tool.poetry]
name = "goit-pythonweb-hw-08"
version = "0.1.0"
description = ""
authors = ["Arnary <anastasia.dushka1@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
sqlalchemy = "^2.0.36"
alembic = "^1.14.0"
asyncpg = "^0.30.0"
fastapi = {extras = ["standart"], version = "^0.115.6"}
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
pydantic-settings = "^2.7.0"
slowapi = "^0.1.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
